Refactor project structure by removing unused data file, updating CSS import paths, and creating new components for About, Blogs, WorkExperience, ProjectList, Contact, Header, and Footer. Update constants and types for better organization and maintainability.
