+++ app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata, Viewport } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import './globals.css'
-import { Header } from './header'
-import { Footer } from './footer'
+import { Header } from '../components/header'
+import { Footer } from '../components/footer'
 import { ThemeProvider } from 'next-themes'
 
 export const viewport: Viewport = {
