+++ components/about.tsx
@@ -1,0 +1,15 @@
+export function About() {
+  return (
+    <div className="flex-1">
+      <p className="text-zinc-600 dark:text-zinc-400">
+        Focused on creating intuitive and performant web experiences. Bridging
+        the gap between design and development.
+      </p>
+      <p className="mt-4 text-zinc-600 italic dark:text-zinc-400">
+        Do not go gentle into that good night.
+        <br />
+        Rage, rage against the dying of the light
+      </p>
+    </div>
+  )
+}
